FIX: Headers bug and logic to use query modality only when embedding videos
